refactor: console 하위 패키지로 분리
